CHAPTER 9 – PRACTICE SET
1. Write a program to read the text from a given file ‘poems.txt’ and find out
whether it contains the word ‘twinkle’.
2. The game() function in a program lets a user play a game and returns the score
as an integer. You need to read a file ‘Hi-score.txt’ which is either blank or
contains the previous Hi-score. You need to write a program to update the Hiscore whenever the game() function breaks the Hi-score.
3. Write a program to generate multiplication tables from 2 to 20 and write it to the
different files. Place these files in a folder for a 13 – year old.
4. A file contains the word “Donkey” multiple times. You need to write a program
which replaces this word with ##### by updating the same file.
5. Repeat program 4 for a list of such words to be censored.
6. Write a program to mine a log file and determine whether it contains ‘python’.
7. Write a program to find out the line number where Python is present from question 6.
8. Write a program to copy a text file “this. txt”
9. Write a program to find out whether a file is identical & matches the content of
Another file.
10. Write a program to wipe out the content of a file using Python.
11. Write a Python program to rename a file to “renamed_by_python.txt.